fix AddressModal2 tests with new relay helpers
